V0.2
Redesigns the cradles to be as minimalistic as possible, moves the
chutes up so they will hopefully catch the coins when they are
released from the cradle, which seems to be at the 70 degree mark.
Thickens the mech top and shortens the pins to allow bolts to hold
the whole thing together.  Also thickens the coin chute and makes
one wall of the chute flat to prevent jams and make it easier to
add coin return later.